refactor(model): replace Integer with BigDecimal for monetary fields
- Producto.precio: Integer -> BigDecimal (precision=10, scale=2)
- DetalleVenta.precio: Integer -> BigDecimal (precision=10, scale=2)
- RequestProductoDto.precio: Integer -> BigDecimal
- DetalleVentaResponseDTO: precioUnitario and subtotal -> BigDecimal
  fix subtotal calculation using .multiply() instead of *
- ProductoService: update price validation using .compareTo(BigDecimal.ZERO)

+++ src/main/java/intisoft2025/practica/dto/producto/RequestProductoDto.java
@@ -5,13 +5,15 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
 public class RequestProductoDto {
     private Long id;
     private String nombre;
-    private Integer precio;
+    private BigDecimal precio;
     private Integer cantidad;
 
     public RequestProductoDto(Producto p) {
